src/test/javascript/spec/app/entities/order-term/order-term.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { OrderTermService } from 'app/entities/order-term/order-term.service';
import { IOrderTerm, OrderTerm } from 'app/shared/model/order-term.model';

describe('Service Tests', () => {
  describe('OrderTerm Service', () => {
    let injector: TestBed;
    let service: OrderTermService;
    let httpMock: HttpTestingController;
    let elemDefault: IOrderTerm;
    let expectedResult: IOrderTerm | IOrderTerm[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(OrderTermService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new OrderTerm(0, 0, 'AAAAAAA', 'AAAAAAA', 0, 0, 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a OrderTerm', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new OrderTerm()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a OrderTerm', () => {
        const returnedFromService = Object.assign(
          {
            sequenceNo: 1,
            name: 'BBBBBB',
            detail: 'BBBBBB',
            termValue: 1,
            termDays: 1,
            textValue: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of OrderTerm', () => {
        const returnedFromService = Object.assign(
          {
            sequenceNo: 1,
            name: 'BBBBBB',
            detail: 'BBBBBB',
            termValue: 1,
            termDays: 1,
            textValue: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a OrderTerm', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
